api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
docker==6.1.3
psutil==5.9.6
netifaces==0.11.0
python-nmap==0.7.1
opencv-python-headless==4.8.1.78
opencv-contrib-python-headless==4.8.1.78
numpy==1.24.4
pillow==10.1.0
aiofiles==23.2.1
python-socketio==5.10.0
celery==5.3.4
flower==2.0.1
